backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
asyncpg==0.27.0
pydantic==1.10.9

# Crypto / Auth
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
